Created a BatchMemberStatus component and used in Header
