+++ packages/nextjs/components/scaffold-eth/index.tsx
@@ -5,3 +5,4 @@ export * from "./Faucet";
 export * from "./FaucetButton";
 export * from "./Input";
 export * from "./RainbowKitCustomConnectButton";
+export * from "./BatchMemberStatus";
